refactor(test): extract scraper_service inline tests to dedicated test file
Move ~959 lines of inline #[cfg(test)] module from scraper_service.rs
to tests/scraper_service_test.rs with explicit imports. Made extract_with_selector,
download_assets_if_enabled, and constants pub for integration test access.
All 45 tests pass in the new location.
